Tests/reorganize: move line length detected tests to own file
